Setup SonarQube analysis
https://sonarcloud.io/organizations/caffeine

+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/policy/product/OhcPolicy.java
@@ -40,7 +40,7 @@ import com.typesafe.config.Config;
  * @author ***REDACTED-EMAIL*** (Ben Manes)
  */
 public final class OhcPolicy implements Policy {
-  private static final int ENTRY_SIZE = 80;
+  private static final long ENTRY_SIZE = 80;
 
   private final OHCache<Long, Long> cache;
   private final PolicyStats policyStats;
